Fix/work around GitHub bug #155.
Here we check that we can get CRT's heap handle because if we do not
then the following `delete` will fail with access violation in
`RtlFreeHeap()`.

How is it possible that the CRT heap handle is NULL?

This function can be called from TLS initializer/terminator by loader
when log4cplus is compiled and linked to as a static library. In case of
other threads temination, it should do its job and free per-thread
data. However, when the whole process is being terminated, it is called
after the CRT has been uninitialized and the CRT heap is not available
any more. In such case, instead of crashing, we just give up and leak
the memory for the short while before the process terminates anyway.

It is possible to work around this situation in user application by
calling `threadCleanup()` manually before `main()` exits.