Added existing function shutdownThreadPool() to the API and added new function restartThreadPool(), which can be called to reset the thread pool after shutting it down through shutdownThreadPool().
This change makes it possible to use log4cplus in applications that use forking (without exec). Since calling fork() doesnt result in threads to be copied over to the child process, and the thread pool is already initialized by default in the parent process, the thread pool in the child process will be in an invalid state as its workers vector is not empty but those worker threads actually dont exist. This bug is especially noticable when trying to exit the child process, the destructor of ThreadPool is then called but waits forever on a conditionto become true, which will never happen. By calling shutdownThreadPool before fork() and then restartThreadPool() right after fork() inside the child and the parent process, both processes can continue using log4cplus correctly.