Support "=" in sub-config file path.
If there is a "=" in the path of sub-config file, and we include it in
the main config file like:

```
include D:/path/to/sub=config.ini
```

it will not work.

Because we parse line with "=" in it as a key/value pair.

This PR is going to fix this issue and support "=" in the "include" directive.

Signed-off-by: Vaclav Haisman <vhaisman@gmail.com>