`filename` should not be empty for TimeBasedRollingFileAppender
TimeBasedRollingFileAppender uses `filenamePattern` alone with
current date time as opened filename and `filename` field is left
empty almost all the time. This brings the problem when it calls
`append()` method from the base class FileAppenderBase.
The base class FileAppenderBase assumes that `filename` field is
the path for opened file. Now this contract is broken.

Currently, this will cause log4cplus to print
"log4cplus: ERROR file is not open:"
when `out` stream is broken. It's confusing and misleading.
Users may think that log4cplus tries to open a file with
the empty name "".

This does not stop here, as long as it persists, more
subtle bugs can be introduced.

This commit fixes this problem by setting filename to the
name of current opened file in
`TimeBasedRollingFileAppender::open()`
making the contract valid again.

Signed-off-by: foxhlchen <foxhlchen@gmail.com>